requirements.txt
attrs==23.1.0
bcrypt==4.0.1
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
dependency-injector==4.41.0
flasgger==0.9.7.1
Flask==2.3.2
Flask-JWT-Extended==4.5.2
flask-marshmallow==0.15.0
Flask-SQLAlchemy==3.0.3
flask-swagger==0.2.14
greenlet==2.0.2
idna==3.4
injector==0.20.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
MarkupSafe==2.1.3
marshmallow==3.19.0
marshmallow-sqlalchemy==0.29.0
mistune==3.0.1
mysqlclient==2.1.1
packaging==23.1
PyJWT==2.7.0
pyrsistent==0.19.3
python-dotenv==1.0.0
PyYAML==6.0
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.16
typing_extensions==4.6.3
urllib3==2.0.3
Werkzeug==2.3.6
